feat(cli): the gantt lanes wear their verbs — and the wire drawing descends to display
lane_chip speaks the 4-verb vocabulary on the flow gantt (accents only · sober bytes intact). wires.rs descends to nika-display behind the decoupled WireGraph (waves of (id,verb) + id edges) — the geometry is now any surface's to draw, the CLI keeps a thin GraphDoc bridge, and nika-cli returns under the 15k wall (the wave had crossed it). Display-side tests are pure constructions (the fixture harness dies).
