+++ crates/nika-display/src/lib.rs
@@ -29,3 +29,4 @@ pub mod snippet;
 pub mod state;
 pub mod theme;
 pub mod vocab;
+pub mod wires;
